gcp_pipeline/cloud_composer/composer_requirements.txt
pandas
pyarrow
google-cloud-storage
google-cloud-bigquery

